Finished refactoring auth logic into auth.js
